+++ src/main/kotlin/hs/kr/equus/feed/infrastructure/kafka/consumer/DeleteFaqTableConsumer.kt
@@ -1,4 +1,4 @@
-package hs.kr.equus.feed.domain.faq.service
+package hs.kr.equus.feed.infrastructure.kafka.consumer
 
 import hs.kr.equus.feed.domain.faq.domain.repository.FaqRepository
 import hs.kr.equus.feed.infrastructure.kafka.configuration.KafkaTopics
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class DeleteFaqTableConsumerService(
+class DeleteFaqTableConsumer(
     private val faqRepository: FaqRepository
 ) {
     @KafkaListener(topics = [KafkaTopics.DELETE_ALL_TABLE], groupId = "delete-all-table-faq")
